chore: move socials into firebase models
