+++ src/models/firebase/socials.d.ts
